Fixed up NuGet paths
The paths in the .csproj file were pointing to fixed paths on the original dev box (E:\CxCache\...) which causes the project not to build on other machines. Fixed up all of these reference paths to point to a local "packages" folder. This is the default for new installs of VS, but for machines that have been configured differently, I added a NuGet.config in the root which will override any machine-specific settings and ensure that NuGet packages get downloaded to the local packages directory.